oreilly_tf/tfagents/requirements.txt
absl-py==2.0.0
ale-py==0.7.5
astunparse==1.6.3
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
blessed==1.20.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==3.0.0
curtsies==0.4.2
cwcwidth==0.1.9
decorator==5.1.1
dm-tree==0.1.8
Farama-Notifications==0.0.4
flatbuffers==23.5.26
gast==0.5.4
gin-config==0.5.0
glfw==2.6.2
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gym==0.23.0
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.10.0
idna==3.4
importlib-resources==6.1.0
keras==2.14.0
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.1.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
Shimmy==0.2.1
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-probability==0.22.1
termcolor==2.3.0
tf-agents-nightly @ file:///home/msmith/opensource/second/agents
tfp-nightly==0.23.0.dev20231027
tqdm==4.66.1
typing_extensions==4.5.0
urllib3==2.0.7
wcwidth==0.2.8
Werkzeug==3.0.1
wrapt==1.14.1
